refactor(ds): extract shared inspector/identityHelpers.tsx (5d.4a)
Pre-step for the MatterInspector + ConversationInspector splits.
Hoists the inspector helper functions that both per-feature components
will need into a single module they can both import, avoiding
duplication.

Extracted to src/shared/ui/inspector/identityHelpers.tsx:
- `InspectorIdentity` type
- `resolveAttorneyLabel(id, matterAssigneeOptions)` — looks up display
  label, falls back to truncated user id
- `resolveAttorneyIdentity(id, conversationMembers, matterAssigneeOptions)`
  — resolves an id to a full identity, looking through conversation
  members first, then assignee options, then synthesizing a fallback
- `renderCompactIdentity(identity)` — small inline avatar + name row
- `renderIdentityStack(users, emptyLabel, singularLabel, pluralLabel)`
  — stacked-avatar row with `{count} {label}` caption

InspectorPanel
--------------
- Imports the extracted helpers
- `resolveAttorneyLabel` / `resolveAttorneyIdentity` thin local wrappers
  via useCallback so dependency arrays stay correct (helpers themselves
  are pure module-level functions now)
- Removed local `renderCompactIdentity` + `renderIdentityStack` useCallbacks
- Removed unused `UserCard` + `StackedAvatars` imports (now used by the
  helper module instead)
- Removed local `InspectorIdentity` type declaration
- Net delete from InspectorPanel: 65 lines

Token cleanups along the way:
- `text-input-placeholder` → `text-dim` in renderIdentityStack
- `text-input-text` → `text-ink` in renderIdentityStack
